Quick and dirty 64bit fix.
`VM_Run` at cg_vm.c:45 will stop executing opcodes until `opPointer` is
`0`. `opPointer` is set in `OP_LEAVE`, stored in the function stack.
`VM_Exec` at cg_vm.c:394 sets this to `0` before calling `VM_Run`. Note,
however, that at cg_vm.c:423 `vm->opPointer` is `0` and
`vm->codeSegment` is some address in memory (e.g. `0x7fd221610010`).
Storing this program counter in an `int32_t` location wraps around.

Thus, after executing the last `OP_LEAVE` opcode, `opPointer` is
`0x7fd400000000`. This is not `0` and the loop will not stop, resulting
in a segmentation fault.

We can *fix* this by casting `onPointer` to an `int32_t`. This is far
from ideal, since it's possible that we break too early.

+++ src/cg_syscall.c
@@ -22,9 +22,9 @@
 
 //handling syscalls from QVM (passing them to engine)
 //this adds the base VM address to a given value
-#define add(x)    ((x) ? (void*)((x) + (int32_t)memoryBase) : NULL)
+#define add(x)    ((x) ? (void*)((x) + (intptr_t)memoryBase) : NULL)
 //this subtracts the base VM address from a given value
-#define sub(x)    ((x) ? (void*)((x) - (int32_t)memoryBase) : NULL)
+#define sub(x)    ((x) ? (void*)((x) - (intptr_t)memoryBase) : NULL)
 //this gets an argument value
 #define arg(x)    (args[(x)])
 //this adds the base VM address to an argument value
